requirements.txt
# ── Core API ───────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9       # file uploads

# ── LLM / LangChain ───────────────────────────────────────────────────────────
openai==1.30.1
langchain==0.2.1
langchain-core==0.2.2
langchain-community==0.2.1
langchain-openai==0.1.8

# ── Vector DB ─────────────────────────────────────────────────────────────────
faiss-cpu==1.13.2
numpy==1.26.4

# ── Settings & validation ─────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.35.0
requests==2.32.2

# ── Dev / test ────────────────────────────────────────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0               # async test client for FastAPI
python-dotenv==1.0.1
